python-version/requirements.txt
requests==2.31.0
google-generativeai==0.3.2
tweepy==4.14.0
schedule==1.2.0
python-dotenv==1.0.0
rich==13.7.0
colorama==0.4.6
